+++ src/main/java/com/online_exam_sys/dao/PaperDao.java
@@ -11,5 +11,5 @@ import com.online_exam_sys.pojo.Paper;
 @Mapper
 public interface PaperDao extends BaseMapper<Paper> {
     @Select("SELECT  * FROM paper WHERE gr_id = (SELECT  gr_id FROM grade_student WHERE st_id = #{st_id})")
-    public List<Paper> queryPaperListByStudentId(Long st_id);
+    public List<Paper> queryPaperListByStudentId(int st_id);
 }
