+++ src/main/java/com/online_exam_sys/dao/StudentDao.java
@@ -9,5 +9,5 @@ import com.online_exam_sys.pojo.Student;
 @Mapper
 public interface StudentDao extends BaseMapper<Student> {
     @Select("SELECT st_password FROM exam_sys.student WHERE st_id = #{id}")
-    public String getPwd(Long id);
+    public String getPwd(int id);
 }
