+++ src/main/java/com/online_exam_sys/pojo/Student.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 @ToString
 public class Student {
     @TableId
-    private Long st_id;
+    private int st_id;
     private String st_name;
     private String st_sex;
     private String st_avatar;
