+++ src/main/java/com/online_exam_sys/service/grade_student/GradeStudentService.java
@@ -1,8 +1,8 @@
 package com.online_exam_sys.service.grade_student;
 
 public interface GradeStudentService {
-    public boolean studentQuitGradeByGsId(Long gs_id);
+    public boolean studentQuitGradeByGsId(int gs_id);
 
-    public boolean studentJoinGradeByGradeId(Long gr_id, Long st_id);
+    public boolean studentJoinGradeByGradeId(int gr_id, int st_id);
 
 }
