+++ src/main/java/com/online_exam_sys/service/teacher/TeacherService.java
@@ -17,5 +17,5 @@ public interface TeacherService {
 
     public Teacher queryByCard(String card);
 
-    public String getPwd(Long id);
+    public String getPwd(int id);
 }
